code/pyproject.toml

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Pedro Farias", email = "pedrohjfg@gmail.com"},
]
dependencies = [
    "fastapi>=0.96.0",
    "uvicorn[standard]>=0.22.0",
    "black>=23.3.0",
    "isort>=5.12.0",
    "pylint>=2.17.4",
    "pydantic>=1.10.9",
    "python-multipart>=0.0.6",
    "SQLAlchemy>=2.0.19",
    "alembic>=1.11.1",
    "python-decouple>=3.8",
    "psycopg2>=2.9.6",
    "bcrypt>=4.0.1",
]
requires-python = ">=3.10"
license = {text = "MIT"}

[tool.isort]
profile = "black"
